+++ compose-action-menu/src/commonMain/kotlin/nl/jacobras/composeactionmenu/OverflowActionItem.kt
@@ -7,9 +7,11 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
+import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -112,5 +114,92 @@ internal fun OverflowActionItem(
                 }
             }
         }
+    )
+}
+
+@PreviewUiComponent
+@Composable
+private fun RegularOverflowActionPreview() {
+    OverflowActionItem(
+        item = RegularActionItem(
+            key = "search",
+            title = "OK",
+            iconVector = Icons.Filled.Search,
+            onClick = {}
+        ),
+        contentColor = MaterialTheme.colorScheme.onSurface
+    )
+}
+
+@PreviewUiComponent
+@Composable
+private fun RegularOverflowActionWithoutIconPreview() {
+    OverflowActionItem(
+        item = RegularActionItem(
+            key = "search",
+            title = "OK",
+            iconVector = null,
+            onClick = {}
+        ),
+        contentColor = MaterialTheme.colorScheme.onSurface
+    )
+}
+
+@PreviewUiComponent
+@Composable
+private fun DisabledOverflowActionWithoutIconPreview() {
+    OverflowActionItem(
+        item = RegularActionItem(
+            key = "search",
+            title = "OK",
+            iconVector = null,
+            enabled = false,
+            onClick = {}
+        ),
+        contentColor = MaterialTheme.colorScheme.onSurface
+    )
+}
+
+@PreviewUiComponent
+@Composable
+private fun CheckableOverflowActionPreview() {
+    OverflowActionItem(
+        item = CheckableActionItem(
+            key = "search",
+            title = "OK",
+            iconVector = Icons.Filled.Search,
+            isChecked = true,
+            onClick = {}
+        ),
+        contentColor = MaterialTheme.colorScheme.onSurface
+    )
+}
+
+@PreviewUiComponent
+@Composable
+private fun RadioOverflowActionPreview() {
+    OverflowActionItem(
+        item = RadioActionItem(
+            key = "search",
+            title = "OK",
+            iconVector = Icons.Filled.Search,
+            isSelected = true,
+            onClick = {}
+        ),
+        contentColor = MaterialTheme.colorScheme.onSurface
+    )
+}
+
+@PreviewUiComponent
+@Composable
+private fun GroupOverflowActionPreview() {
+    OverflowActionItem(
+        item = GroupActionItem(
+            key = "search",
+            title = "OK",
+            iconVector = Icons.Filled.Search,
+            childOptions = emptyList()
+        ),
+        contentColor = MaterialTheme.colorScheme.onSurface
     )
 }
