+++ compose-action-menu/src/commonMain/kotlin/nl/jacobras/composeactionmenu/PreviewAnnotations.kt
@@ -1,7 +1,7 @@
 package nl.jacobras.composeactionmenu
 
-import android.content.res.Configuration.UI_MODE_NIGHT_YES
-import android.content.res.Configuration.UI_MODE_TYPE_NORMAL
+import androidx.compose.ui.tooling.preview.AndroidUiModes.UI_MODE_NIGHT_YES
+import androidx.compose.ui.tooling.preview.AndroidUiModes.UI_MODE_TYPE_NORMAL
 import androidx.compose.ui.tooling.preview.Preview
 
 /**
